+++ app/src/main/java/woowacourse/movie/domain/model/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model.data
+package woowacourse.movie.domain.model
 
 import java.io.Serializable
 import java.time.LocalDate
