+++ app/src/test/java/woowacourse/movie/model/MovieTicketTest.kt
@@ -1,11 +1,11 @@
-package woowacourse.movie.data
+package woowacourse.movie.model
 
 import io.kotest.assertions.assertSoftly
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.movie.model.data.MovieTicket
-import woowacourse.movie.model.policy.DefaultPricingPolicy
+import woowacourse.movie.domain.model.MovieTicket
+import woowacourse.movie.domain.policy.DefaultPricingPolicy
 import java.time.LocalDateTime
 
 class MovieTicketTest {
