+++ app/src/test/java/woowacourse/movie/schedule/MovieSchedulerTest.kt
@@ -6,7 +6,7 @@ import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.movie.model.schedule.MovieScheduler
+import woowacourse.movie.domain.schedule.MovieScheduler
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
